Add init.py to initialize utils/ and move append_hostname there
